Reuse landing page for dashboard
