+++ packages/extension/src/view/devtools/components/privacySandbox/index.tsx
@@ -17,18 +17,14 @@
  * External dependencies.
  */
 import React, { useEffect } from 'react';
-import {
-  QuickLinksList,
-  PrivacySandboxColoredIcon,
-  ExternalLinkBlack,
-  useSidebar,
-  SIDEBAR_ITEMS_KEYS,
-} from '@google-psat/design-system';
-import { addUTMParams } from '@google-psat/common';
+import { LandingPage } from '@google-psat/design-system';
+
+/**
+ * Internal dependencies.
+ */
+import ContentPanel from './contentPanel';
 
 const PrivacySandbox = () => {
-  const navigateTo = useSidebar(({ actions }) => actions.updateSelectedItemKey);
-
   useEffect(() => {
     (async () => {
       await chrome.storage.sync.set({
@@ -38,63 +34,7 @@ const PrivacySandbox = () => {
   }, []);
 
   return (
-    <div
-      data-testid="privacy-sandbox-content"
-      className="h-full w-full overflow-auto"
-    >
-      <div className="p-4 dark:text-bright-gray">
-        <header className="flex items-center">
-          <PrivacySandboxColoredIcon
-            width="25"
-            height="25"
-            className="mr-1.5"
-          />
-          <h1 className="text-lg">Privacy Sandbox</h1>
-        </header>
-        <section className="flex justify-center pt-3">
-          <div className="max-w-screen-md text-center">
-            <PrivacySandboxColoredIcon
-              width="90"
-              height="90"
-              className="inline-block mb-5"
-            />
-            <h2 className="text-5xl mb-5 font-semibold">
-              Protecting your privacy online
-            </h2>
-            <p className="text-base">
-              The Privacy Sandbox initiative aims to create technologies that
-              both protect people&apos;s privacy online and give companies and
-              developers tools to build thriving digital businesses. The Privacy
-              Sandbox reduces cross-site and cross-app tracking while helping to
-              keep online content and services free for all.
-            </p>
-            <div className="flex gap-6 justify-center mt-5">
-              <a
-                href={addUTMParams('https://privacysandbox.com')}
-                target="__blank"
-                className="bg-cultured-grey text-raisin-black py-2 px-9 rounded border border-dark-grey text-base hover:bg-light-gray hover:border-american-silver flex"
-              >
-                <span>Learn About Privacy Sandbox</span>
-                <ExternalLinkBlack
-                  width="16"
-                  height="16"
-                  className="mt-1 ml-1"
-                />
-              </a>
-              <button
-                onClick={() => navigateTo(SIDEBAR_ITEMS_KEYS.DASHBOARD)}
-                className="bg-cultured-grey text-raisin-black py-2 px-9 rounded border border-dark-grey text-base hover:bg-light-gray hover:border-american-silver"
-              >
-                Dashboard
-              </button>
-            </div>
-          </div>
-        </section>
-      </div>
-      <div className="mt-10">
-        <QuickLinksList />
-      </div>
-    </div>
+    <LandingPage title="Privacy Sandbox" contentPanel={<ContentPanel />} />
   );
 };
